move editor app to pages
