--- src/editor/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Editor</title>
-    <link rel="stylesheet" href="editor.scss">
-</head>
-<body>
-    <div id="app"></div>
-    <script type="module" src="editor.tsx"></script>
-</body>
-</html>
