+++ src/pages/editor/editor.scss
