+++ src/pages/editor/types.ts
